Modificando carpetas para que siga Clean Arquitechture
