+++ sprint 2/BackendGeems/BackendGeems/Application/IGEEMSRepo.cs
@@ -1,0 +1,10 @@
+﻿using BackendGeems.Domain;
+namespace BackendGeems.Application
+{
+    public interface IGEEMSRepo
+    {
+        bool calcularPago(string fechaInicio, string fechaFinal);
+        List<Pago> ObtenerPagos(DateTime fechaInicio, DateTime fechaFinal);
+        int ObtenerSalarioBruto(Guid idEmpleado, DateTime fechaInicio, DateTime fechaFinal);
+    }
+}
