--- sprint 2/BackendGeems/BackendGeems/Application/QueryPago.cs
@@ -1,21 +1,0 @@
-﻿using BackendGeems.Domain;
-using BackendGeems.Infraestructure;
-
-namespace BackendGeems.Application
-{
-    public class QueryPago : IQueryPago
-    {
-        private readonly IGEEMSRepo _repoInfrastructure;
-        public QueryPago(IGEEMSRepo repo)
-        {
-            _repoInfrastructure = repo;
-        }
-        public List<Pago> ObtenerPagos(DateTime fechaInicio, DateTime fechaFinal)
-        {
-           var pagos = _repoInfrastructure.ObtenerPagos(fechaInicio, fechaFinal);
-            return pagos;
-        }
-    }
-    
-}
-
