+++ sprint 2/BackendGeems/BackendGeems/Controllers/RegisterController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.Data.SqlClient;
-using BackendGeems.Models;
+using BackendGeems.Domain;
 
 namespace BackendGeems.Controllers
 {
