+++ sprint 2/BackendGeems/BackendGeems/Domain/Beneficio.cs
@@ -1,0 +1,11 @@
+namespace BackendGeems.Domain
+{
+    public class Beneficio
+    {
+        public required string Nombre { get; set; }
+        public required string Descripcion { get; set; }
+        public int Costo { get; set; }
+        public int TiempoMinimo { get; set; }
+        public required string CedulaJuridica { get; set; }
+    }
+}
