+++ sprint 2/BackendGeems/BackendGeems/Infraestructure/GEEMSRepo.cs
@@ -1,4 +1,5 @@
-﻿using BackendGeems.Domain;
+﻿using BackendGeems.Application;
+using BackendGeems.Domain;
 using Microsoft.Data.SqlClient;
 using System.Data;
 using System.Data.SqlClient;
