--- sprint 2/BackendGeems/BackendGeems/Models/Beneficio.cs
@@ -1,8 +1,0 @@
-public class Beneficio
-{
-    public required string Nombre { get; set; }
-    public required string Descripcion { get; set; }
-    public int Costo { get; set; }
-    public int TiempoMinimo { get; set; }
-    public required string CedulaJuridica { get; set; }
-}
